requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-core==1.35.1
azure-storage-blob==12.26.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.1
et_xmlfile==2.0.0
fastapi==0.116.2
h11==0.16.0
html2docx==1.6.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
openpyxl==3.1.5
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tinycss2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1